pyproject.toml
[tool.poetry]
name = "football"
version = "0.1.0"
description = ""
authors = ["Evgeny Dyshlyuk <evgeny.dyshlyuk@gmail.com>"]
readme = "README.md"

packages = [
  { include = "app" }
]

[tool.poetry.dependencies]
python = "3.13.3"
fastapi = "0.115.14"
uvicorn = { version = "0.35.0", extras = ["standard"] }
jinja2 = "3.1.6"
pydantic = "2.11.7"
python-multipart = "0.0.20"
python-jose = "^3.5.0"
boto3 = "^1.39.3"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
